requirements.txt
apache-airflow==2.1.0
apache-airflow-providers-postgres==1.0.2
slack-sdk==3.6.0